+++ packages/components/src/components/input-email/input-email-utils.ts
@@ -1,5 +1,7 @@
+import type { FormFieldAriaAttribute } from '../../types';
 import type { FormState } from '../../utils/form/form-state';
 export type InputEmailState = FormState;
+export type InputEmailAriaAttribute = FormFieldAriaAttribute;
 
 export type InputEmailChangeEventDetail = Event;
 export type InputEmailBlurEventDetail = Event;