+++ packages/components/src/components/input-url/input-url-utils.ts
@@ -1,5 +1,7 @@
+import type { FormFieldAriaAttribute } from '../../types';
 import type { FormState } from '../../utils/form/form-state';
 export type InputUrlState = FormState;
+export type InputUrlAriaAttribute = FormFieldAriaAttribute;
 
 export type InputUrlChangeEventDetail = Event;
 export type InputUrlBlurEventDetail = Event;