+++ packages/components/src/types/common-types.d.ts
@@ -3,7 +3,7 @@ import type { IconName } from '@porsche-design-system/icons';
 
 export type { FlagName } from '@porsche-design-system/flags';
 export type { IconName } from '@porsche-design-system/icons';
-export type { ButtonAriaAttribute, LinkAriaAttribute } from '../utils';
+export type { ButtonAriaAttribute, FormFieldAriaAttribute, LinkAriaAttribute } from '../utils';
 export type { BreakpointCustomizable } from '../utils/breakpoint-customizable';
 export type { PropTypes, ValidatorFunction } from '../utils/validation/validateProps';