+++ packages/components/src/utils/a11y/index.ts
@@ -1,4 +1,5 @@
 export * from './a11y';
 export * from './button';
+export * from './form-field-aria-attribute';
 export * from './link';
 export * from './select';