+++ packages/storefront/src/app/(main)/components/switch/switch.meta.ts
@@ -1,6 +1,7 @@
 import { componentMeta } from '@porsche-design-system/component-meta';
 import { hiddenLabelWithoutAccessibleNameA11yExample } from '@/app/(main)/components/switch/accessibility/examples/hidden-label-without-accessible-name/example';
 import { switchWithoutDescriptiveLabelA11yExample } from '@/app/(main)/components/switch/accessibility/examples/switch-without-descriptive-label/example';
+import { ariaOnComponentHostVsAriaPropA11yExample } from '@/app/(main)/components/switch/accessibility/examples/aria-on-component-host-vs-aria-prop/example';
 import AccessibilityOverview from '@/app/(main)/components/switch/accessibility/overview.mdx';
 import AccessibilityTests from '@/app/(main)/components/switch/accessibility/tests.mdx';
 import IntroductionDescription from '@/app/(main)/components/switch/configurator/introduction.mdx';
@@ -18,6 +19,7 @@ export const switchMeta = {
   accessibility: {
     overview: AccessibilityOverview,
     examples: {
+      ariaOnComponentHostVsAriaProp: ariaOnComponentHostVsAriaPropA11yExample,
       switchWithoutDescriptiveLabel: switchWithoutDescriptiveLabelA11yExample,
       hiddenLabelWithoutAccessibleName: hiddenLabelWithoutAccessibleNameA11yExample,
     },